Pequenos ajustes para manter consistência
