+++ src/routes/UserRouter.ts
@@ -1,0 +1,3 @@
+import express from 'express';
+
+export const userRouter = express.Router();
